requirements.txt
# Core ML/DL Libraries
tensorflow==2.14.0
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
Pillow==10.1.0

# Experiment Tracking & Versioning
mlflow==2.10.2
dvc==3.47.0
dvc-s3==3.3.0

# Model Serving & API
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Testing & Code Quality
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
